requirements.txt
streamlit
firebase-admin
google-cloud-firestore
python-dotenv
langchain
langchain-groq